+++ src/components/DraggableList.module.css
@@ -1,0 +1,5 @@
+.draggableDailogWrapper {
+  background-color: lightgrey;
+  box-shadow: 5px 5px 5px rgb(112, 112, 112);
+  margin-bottom: 5px;
+}
