--- src/components/layout.js
@@ -1,9 +1,0 @@
-import styles from "./layout.module.css"
-
-export const Layout = ({ children, home }) => {
-    return (
-        <div className={styles.container} >
-            <main>{children}</main>
-        </div>
-    )
-}
