+++ src/components/layout.module.css
@@ -1,5 +1,25 @@
 .container {
-    margin: auto;
-    width: 30rem;
-    text-align: center;
+  margin: auto;
+  width: 100%;
+  height: 100vh;
+  text-align: center;
+  background-color: rgba(235, 246, 235, 0.568);
 }
+
+.header {
+  background-color: black;
+  color: white;
+  padding: 16px;
+}
+.main {
+  padding: 16px;
+  height: 100%;
+}
+/* @media screen and (max-width: 480px) {
+  .container {
+  }
+  .header {
+  }
+  .main {
+  }
+} */
